@@ -67,19 +67,18 @@
     "mobile-detect": "^1.4.4",
     "net": "^1.0.2",
     "network-speed": "^2.1.1",
-    "next": "^10.0.8",
+    "next": "^11.1.2",
     "next-i18next": "^8.4.0",
-    "next-images": "^1.2.0",
     "next-plugin-transpile-modules": "^2.0.0",
     "node-cache": "^5.0.2",
     "nuka-carousel": "^4.5.13",
     "raw-loader": "^4.0.1",
-    "react": "^17.0.1",
+    "react": "^17.0.2",
     "react-apollo": "^3.1.3",
     "react-async-hook": "^3.6.1",
     "react-autosuggest": "^10.0.2",
     "react-awesome-reveal": "^3.3.1",
-    "react-dom": "^17.0.1",
+    "react-dom": "^17.0.2",
     "react-ga": "^3.1.2",
     "react-google-recaptcha": "^2.0.1",
     "react-hook-form": "^7.1.1",
@@ -144,7 +143,7 @@
     "ts-jest": "^26.5.3",
     "ts-node": "^9.1.1",
     "tsconfig-paths": "^3.8.0",
-    "typescript": "^4.2.3"
+    "typescript": "^4.4.3"
   },
   "lint-staged": {
     "*.{ts, tsx}": [
