@@ -1,5 +1,5 @@
 import * as React from "react"
-import { Image, ImageSourcePropType, ImageStyle, StyleSheet, Text, View } from "react-native"
+import { Image, ImageStyle, StyleSheet, Text, View } from "react-native"
 import { H3 } from "src/fonts/Fonts"
 import { Cell, Spans } from "src/layout/GridRow"
 import { useScreenSize } from "src/layout/ScreenSize"
@@ -7,7 +7,7 @@ import Button, { BTN, SIZE } from "src/shared/Button.3"
 import { fonts, standardStyles } from "src/styles"
 
 interface Props {
-  source: ImageSourcePropType
+  source: string | StaticImageData
   title: string
   text: string
   link?: {
@@ -24,7 +24,7 @@ export function Adventure({ title, text, source, link, imageStyle }: Props) {
       style={[styles.root, isMobile && standardStyles.elementalMarginBottom]}
     >
       <View>
-        <Image source={source} style={[styles.image, imageStyle]} resizeMode="contain" />
+        <Image source={{uri: typeof source === "string" ? source : source.src}} style={[styles.image, imageStyle]} resizeMode="contain" />
         <H3 style={standardStyles.elementalMargin}>{title}</H3>
         <Text style={fonts.p}>{text}</Text>
       </View>
