@@ -1,60 +1,53 @@
 import * as React from "react"
-import { ImageBackground, ImageSourcePropType, StyleSheet, Text } from "react-native"
-import { useScreenSize } from "src/layout/ScreenSize"
-import { colors, fonts, standardStyles, textStyles } from "src/styles"
+import { ImageBackground, StyleSheet } from "react-native"
+import { colors,  standardStyles } from "src/styles"
+import * as eStyles from "src/estyles"
+import { css } from "@emotion/react"
 
 interface Props {
-  imgSource: ImageSourcePropType
+  imgSource: StaticImageData
   quote: React.ReactNode
   citation: React.ReactNode
   color?: colors
 }
 
 export default React.memo(function BeautifulQuote(props: Props) {
-  const { isMobile } = useScreenSize()
   return (
     <ImageBackground
-      source={props.imgSource}
+      source={{uri:props.imgSource.src}}
       style={[styles.image, standardStyles.centered]}
       resizeMode={"cover"}
     >
-      <Text
-        style={[
-          fonts.h1,
-          isMobile ? styles.quoteMobile : styles.quote,
-          textStyles.invert,
-          textStyles.center,
-          props.color && { color: props.color },
-        ]}
-      >
+      <blockquote css={css(quoteCss,props.color && { color: props.color })} >
         “{props.quote}”
-      </Text>
-      <Text
-        style={[
-          isMobile ? fonts.h1Mobile : fonts.h1,
-          textStyles.invert,
-          textStyles.center,
-          standardStyles.blockMarginTopTablet,
-          props.color && { color: props.color },
-        ]}
-      >
+      </blockquote>
+      <figcaption css={css(citeCss, props.color && { color: props.color })} >
         {props.citation}
-      </Text>
+      </figcaption>
     </ImageBackground>
   )
 })
 
+const quoteCss = css(eStyles.fonts.h1,
+  eStyles.textStyles.center,
+  {
+  fontSize: 65,
+  lineHeight: "72px",
+  fontStyle: "italic",
+  maxWidth: 1200,
+  [eStyles.WHEN_MOBILE]: {
+    fontSize: 42,
+    lineHeight: "50px",
+    fontStyle: "italic",
+  }
+})
+
+const citeCss = css(eStyles.fonts.h1,  eStyles.textStyles.center, {
+  marginTop: 40,
+  fontSize: 36,
+  [eStyles.WHEN_MOBILE]: eStyles.fonts.h1Mobile,
+})
+
 const styles = StyleSheet.create({
   image: { width: "100%", minHeight: 510, paddingVertical: 15, paddingHorizontal: 30 },
-  quote: {
-    fontSize: 65,
-    lineHeight: 72,
-    fontStyle: "italic",
-    maxWidth: 1200,
-  },
-  quoteMobile: {
-    fontSize: 42,
-    lineHeight: 50,
-    fontStyle: "italic",
-  },
 })
